llm/chatglm3/pyproject.toml
[project]
name = "chatglm3"
version = "0.1.0"
description = "https://github.com/THUDM/ChatGLM3"
requires-python = ">=3.10"
dependencies = [
    "cpm-kernels>=1.0.11",
    "flask>=3.1.2",
    "loguru>=0.7.3",
    "numpy==1.26.4",
    "sentencepiece==0.2.0",
    "setuptools>=80.9.0",
    "torch==2.3.1",
    "transformers==4.41.2",
    "modelscope==1.33.0",
    "zerolan-data",
]

[tool.uv.sources]
zerolan-data = { git = "https://github.com/AkagawaTsurunaki/zerolan-data.git", rev = "1.5.0" }

[tool.setuptools]
packages = []
